refract contrib polaris client dia option
